environment/modules/aws/lambda/functions/sqs-custom-metrics-per-minute/requirements.txt
awscli==1.18.36
python-dateutil==2.8.0